Move cache options into controller package
